requirements.txt
# API Server
fastapi==0.109.0
uvicorn==0.27.0
python-multipart==0.0.6

# Database
aiosqlite==0.19.0

# Environment
python-dotenv==1.0.0
